# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Event-Helper",
    "description": "Event managment and handler web app"
  },
  "host": "https://event-helper-server.herokuapp.com",
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/register": {
      "put": {
        "tags": ["User"],
        "summary": "Register an account",
        "description": "",
        "operationId": "registration",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "User params",
            "name": "User params",
            "description": "User must provide email, password and name",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userName": {
          "type": "string"
        },
        "userEmail": {
          "type": "string"
        },
        "userPassword": {
          "type": "string"
        },
        "isAdmin": {
          "type": "boolean"
        }
      }
    }
  }
}